+++ src/components/form/input-base/types.ts
@@ -1,0 +1,6 @@
+export interface InputBaseProps
+  extends React.InputHTMLAttributes<HTMLInputElement> {
+  fieldSetProps?: React.HTMLAttributes<HTMLElement>;
+  label?: string;
+  labelProps?: React.HTMLAttributes<HTMLLabelElement>;
+}